feat(ui): refactor wallet modal interfaces and expose them publicly
