+++ packages/ui/src/app/state/modals-state.ts
@@ -2,6 +2,7 @@ import { createMemo, createSignal } from 'solid-js';
 import { WalletInfoWithOpenMethod, WalletOpenMethod } from 'src/models/connected-wallet';
 import { LastSelectedWalletInfoStorage } from 'src/storage';
 import { ReturnStrategy } from 'src/models';
+import { WalletsModalState } from 'src/models/wallets-modal';
 
 export type ActionName = 'confirm-transaction' | 'transaction-sent' | 'transaction-canceled';
 
@@ -18,46 +19,6 @@ export type ConfirmTransactionAction = BasicAction & {
     returnStrategy: ReturnStrategy;
     twaReturnUrl: `${string}://${string}`;
 };
-
-/**
- * Opened modal window state.
- */
-export type WalletModalOpened = {
-    /**
-     * Modal window status.
-     */
-    status: 'opened';
-
-    /**
-     * Always `null` for opened modal window.
-     */
-    closeReason: null;
-};
-
-/**
- * Closed modal window state.
- */
-export type WalletModalClosed = {
-    /**
-     * Modal window status.
-     */
-    status: 'closed';
-
-    /**
-     * Close reason, if the modal window was closed.
-     */
-    closeReason: WalletsModalCloseReason | null;
-};
-
-/**
- * Modal window state.
- */
-export type WalletsModalState = WalletModalOpened | WalletModalClosed;
-
-/**
- * Modal window close reason.
- */
-export type WalletsModalCloseReason = 'action-cancelled' | 'wallet-selected';
 
 export const [walletsModalState, setWalletsModalState] = createSignal<WalletsModalState>({
     status: 'closed',
